Let constructor stubs always throw
Summary:
Consider the following code:
```
class A {
  public function __construct(private int $i) {}
}
class B extends A {
  public function __construct(private string $s) {
    parent::__construct(0);
  }
}
function target(): void {
  $a = new A(42);
  $b = new B('foo');
}
```
At the moment, extraction of `target` will pull in the dependencies on `A` and `B` and their constructors (because instances of `A` and `B` are created by `target`), but the generated constructors will have empty bodies and in particular the extracted constructor of `B` will not be calling `parent::__constructor`, which it has to do to initialize the property `$this->i` inherited from `A`.  The extracted code produces an error
```
NastCheck[3012] You are extending a class that needs to be initialized
Make sure you call parent::__construct.
```

This problem can be easily solved by making constructor bodies to be `throw new Exception();` because in this case the type checker is smart enough to figure out that the constructor doesn't return and hence there in no point in checking all properties are initialized.  This also simplifies code in the case when there are properties that need to be initialized and that are currently initialized by calling `extract_standalone_make_default`.

Reviewed By: kmeht

Differential Revision: D17879109

fbshipit-source-id: dd320a949378c69ab8b826bdf82927870777dbae