Fix tag_provenance_here to actually CoW
Summary:
This is almost invariably called on static arrays and it would be the
Wrong Thing(TM) to tag the input array without copying it.

This is causing us to inadvertantly share provenance information between e.g.
distinct calls to Vec\map and friends and is Not A Good Thing(TM)

Reviewed By: billf

Differential Revision: D17856041

fbshipit-source-id: 779d898038463d5024b6890379a0a28b922b5409