Switch AutocompleteService to always use sienv
Summary:
Back before we shipped new autocomplete, I created lots of special code pathways to ensure that I could work on the project without affecting users before I was ready to begin delivering new autocomplete.  Now that we've shipped new autocomplete and it's been stable for a few months, let's remove the old code.

This particular diff updates the normal autocomplete service to use the normal symbol index pathway for all global searches.

Because this is the file that had the most customized code, this diff removes a lot of old stuff.

Reviewed By: arxanas

Differential Revision: D17863989

fbshipit-source-id: c2fa4431f4e42583acdd0aa8030834e301c95eae