Fix foreach statements involving dynamic
Summary:
Currently the typing of dynamic and foreach statements is incorrect. Consider the following code:

```
class C {}

function bad_foreach(dynamic $dyn, C $c): void {
  if ($dyn) {
    $list = $dyn;
  } else {
    $list = $c;
  }
  foreach ($list as $x) {}
}
```

Currently this code will type check, because we allow to foreach over a value if dynamic is a subtype of it. This is incorrect and we should report an error because `C` is not a `Traversable`.

At the same time if you foreach over a `~Traversable<T>` we want `$x` to have the type `~T`.

This diff fixes this (as an alternative to {D17801884}), by distributing across unions and pattern matching on dynamic in Typing.as_expr

Reviewed By: andrewjkennedy

Differential Revision: D17880800

fbshipit-source-id: 225fed274eb6828cc99db67e3eeb1803e32ffc2d