Deprecate error 2053 (referencing HH\BuiltinEnum outside of an hhi)
Summary: `BuiltinEnum` was originally meant to be an implementation detail of enums, where we provided some functionality to each of them. People though have used it a bit to determine if something is an enum at all since we didn't provide a great way to determine that. Since we still don't have a great way to determine that, we're just going to deprecate this error.

Reviewed By: kmeht

Differential Revision: D17911073

fbshipit-source-id: 9f7df2bbff2ba9e13aad56a72cf4859c31712efa