Do not pessimize return types
Summary:
As a consequence of our new coercion algorithm, it is no longer necessary to pessimize types on writes, i.e. as we're exiting the scope of a function. With `dynamic` treated as bottom type in coercion situations, it is not necessary to pessimize the type on the right for `~shape('a'=> int) ~> shape('a' => int)` to succeed.

This change thereby prevents like types from being exposed as the `Expected` type in error messages.

Reviewed By: dlreeves

Differential Revision: D17727042

fbshipit-source-id: e7b18c2dbbf3b01007868a5a8f93744e64e3a9a4