Fix integration tests for status
Summary:
These tests were broken/flaky because we always `sleep 1.0`'d before emitting a status, so we never saw the "initializing" status message. Moving the `sleep` afterward and showing status immediately ensures that we don't miss any status messages.

To fix this, we also have to make it possible for the tests to receive a `showStatus` request and not respond to it. (If we do respond to it, then the language server immediately sends another `showStatus` request, until you don't respond to it.)

Reviewed By: jewelpit

Differential Revision: D17938029

fbshipit-source-id: b1834cf1369ce7aa9d95b0dfeaa3f8254cb99a91