Implement iter specialization IR ops
Summary:
Here are the IR ops we use in iterator specialization. There are 17 new IR op codes in this diff, so I'll summarize them.

First, there are 9 ops for accessing iterator fields. These are all quite simple. We split them up so that we can give them discrete memory effects, because we can usually load- and store-elim them away:
- **StIter{Base,Type,Pos,End}:** IR setters for iterator fields - simple PureStore memory ops. StIterType has special load-elim handling effects because it writes an immediate to the type field.
- **LdIter{Base,Pos,End} + CheckIter:** IR getters for iterator fields. "pos" and "end" take a type param: an int (for index iteration) or a subtype of PtrToElemGen (for pointer iteration).
- **KillIter:** helps with debugging and causes store-elim to kick in whenever load-elim does.

Second, there are 7 ops specific to pointer iteration on PackedArray and MixedArray elements:
- **Get{Mixed,Packed}PtrIter:** returns a pointer iterator at an index in an array with a given layout.
- **Advance{Mixed,Packed}PtrIter:** increments or decrements a pointer iterator.
- **LdPtrIter{Key,Val}:** loads a pointer iterator's key or value. The key getter takes a type param.
- **EqPtrIter:** compare pointer iterators for equality.

Finally, for loops over dicts, we want to check the key types before jumping to specialized code. I think this op could be useful for other optimizations:
- **CheckMixedArrayKeys<T>:** an op that uses a one-byte test to verify that all keys of a MixedArray have a given type. Valid T include Int, Str, StaticStr, and Int|Str. This check may have false negatives, like CheckMixedArrayOffset.

Reviewed By: mofarrell

Differential Revision: D17824977

fbshipit-source-id: 87bd05d13919bf4f3f54c6e2fa5058856f701cc5