Analyse scope of pipe variable in typing, not naming
Summary:
The naming phase now does very little scope analysis, using the integer first component of a `Local_id.t` only for (a) the experimental let-bound variables, and (b) pipe variables `$$`.

But these latter variables also don't need the integer: it's easy to scope them properly during typing, with the expression `e1 |> e2` typed by adding `$$` of type `t` (where `t` is the type of `e1`) to the environment when typing `e2`, and then restoring `$$` (or unsetting it) to its original state.

We also need to remove the variable under a lambda, as this is not supported by HackC.

Reviewed By: CatherineGasnier

Differential Revision: D17880302

fbshipit-source-id: 2e28d3909e339e52dbc5a094c16d185489857138