Clean up some collection raw strings in emitter
Summary:
The emitter has a bad habit of passing around and comparing raw strings directly (as opposed to referring to the ones listed in Naming_special_names).

This causes a monumental headache when trying to unify the namespace elaboration behavior between the typechecker and the emitter. So I'll take a quick pass over the emitter and try and remove some of that behavior.

This diff addresses some bad string behavior with regards to Collections (`Vector`, `ImmMap` and friends). Trying to (instead of passing around raw strings) turn them into `CollectionType.t` as quickly as possible. And also refactored some emission code.

Reviewed By: losvald

Differential Revision: D17944223

fbshipit-source-id: f7df4dc7fba4c3c2d9a9537c7988a2d06f13a920