Do not reformat function declaration headers preceded by a FIXME
Summary:
Somehow, the logic for handling a FunctionDeclarationHeader was duplicated into MethodishDeclaration (and was then copy-pasted into MethodishTraitResolution).

Duplicating the transform logic rather than invoking the transform function (which is named `t`) circumvents our formatting-suppression handling (on the first couple lines of `t`). The formatting rules for FunctionDeclarationHeader don't need to be duplicated in the first place, and using the normal path allows FIXME-suppression to work properly on methods.

Reviewed By: DavidSnider

Differential Revision: D17953613

fbshipit-source-id: 5425051620be42cfeb70846c77deeb43b2d084d5