Use coercion to type arithmetic/bitwise operators
Summary: Due to usage of `Typing_utils.is_dynamic` we would allow ~T to be valid for all arithmetic and bitwise operators even if T was not a num or int. I fix this by utilizing coercion. We know that the runtime will enforce that the arithmetic/bitwise operators are a num/int so we can coerce to enforceable num/int types.

Reviewed By: vassilmladenov

Differential Revision: D17946429

fbshipit-source-id: 6566f0c18770b0c9801bec039a7ea85fdabccbd5