Fix RegionDesc and CallTargets TC dump annotations
Summary:
D16838354 added annotations to IRUnit, move inlining decisions there, and
stopped including the annotations from IRGS in the TC dump.  This resulted in
the other annotations that were added to IRGS to be lost, namely RegionDesc and
CallTargets annotations.  This diff moves those annotations to also use IRUnit's
annotationData and drops the annotations from IRGS.

Reviewed By: jjgriego

Differential Revision: D17949471

fbshipit-source-id: c3ba48e028a5d51196f2d76d4d46fc02df2b5b72