Model namespace aliasing as implicit 'use namespace' statements
Summary:
As a result of past work to standardize namespace aliasing between the typechecker and compiler, namespace aliasing is semantically equivalent to a series of `use namespace` statements at the top of every file and/or namespace block (except that they can be overridden by users in code).

This diff literally considers aliases to be `use namespace` statements, allowing us to kill all special-cased code for namespace aliasing and instead re-use the existing elaboration machinery. Spoiler Alert: when HH-autoimporting is fully unified, we can do the same thing for it!

Reviewed By: Matt-Schellhas

Differential Revision: D17817943

fbshipit-source-id: f93903ad204e7f3b00f258f4664af7f44772c889