Do not pessimize function argument types
Summary:
Similar to {D17727042}, this removes pessimization of function arguments in a call.

Previously, the model was that if a type is unenforceable, then it is transformed to a like type at coercion points. Therefore, the code
```
function expect<T>(T $t): void {
  hh_show($t); // ~T
}

function caller(dynamic $d, string $s): void {
  expect<int>($d); // expect<int>: (function (~int): void)
  // call succeeds
  //    because dynamic ~> (dynamic | int)
  //    because dynamic <: (dynamic | int)

  expect<int>($s); // error, expected ~int, got string.
}
```
The new model is that we only transform an unenforceable type to a like type when we "read" it, as in the body of the `expect` function. When we "write" it, as in the call to `expect` in `caller`, we do not need to pessimize it, because unenforceable types are valid coercion targets if pessimization is enabled. Instead, from {D17682757} we treat `dynamic` as a bottom type, so the following behavior is observed:

```
function expect<T>(T $t): void {
  hh_show($t); // ~T
}

function caller(dynamic $d, string $s): void {
  expect<int>($d); // expect<int>: (function (int): void)
  // call succeeds
  //    because dynamic ~> int (unenforced)
  //    because dynamic <: int
  //    because dynamic is treated as a bottom type at coercion points

  expect<int>($s); // error, expected int, got string.
}
```

The benefit to this approach is that the typing errors largely remain the same as the ones from before pessimization.

Finally, we have some parameters that are both "write" and "read" in the form of `inout` parameters. Since we are reading, we need to pessimize inout parameters that are unenforced.

```
function f<T>(inout T $tout, dynamic $d): void {
  hh_show($tout); // ~T
  $tout = $d; // dynamic ~> ~T
}

function caller2(int $i): void {
  f(inout $i, "hello");
  // We read $i from an unenforced inout argument, so it is a ~int.
  hh_show($i); // ~int
  var_dump($i); // "hello";
}

Reviewed By: dlreeves

Differential Revision: D17726684

fbshipit-source-id: 0ab2178dff81824fa44ea4dcb4afc56800a0963e