Stop allowing autoimported classes to be overridden
Summary:
We had a special case built into HH-autoimporting that allowed us to override it if we defined a class with the same name as an autoimported class within a file, allowing us to refer to it plainly (instead of having to use a `namespace\` prefix. However, this only worked for classes (not functions or consts), and only worked for classes defined in the same file!

It didn't seem like a feature worth complicating our already bizarre autoimporting logic; this diff removes it.

```
namespace Foo;

class Vector {}

function f() {
  // Before: Foo\Vector
  // After: HH\Vector
  new Vector();
}
```

Reviewed By: Matt-Schellhas

Differential Revision: D17722540

fbshipit-source-id: 21c87ad3ae1298bac5b4e2e5e88c26148634266e