Drop kHasClassBit from ActRec
Summary: In lowptr builds use the fact that Class* is in lowmem and ObjectData* is not to distinguish between Class/Object in m_thisUnsafe. We need to do this to avoid a dcache regression in runtime code that still has to inspect whether the frame contains a class or an object.

Reviewed By: jano

Differential Revision: D16119541

fbshipit-source-id: be869b3a1512192e05e9b096876d26bc764373b3