Better way of detecting built-in dependencies
Summary: At the moment, to determine if a given dependency is built-in (and hence its declaration must not be generated in the extracted code), we check if the origin of the dependency is a file with the `.hhi` extension.  This doesn't work for custom `.hhi` files that may be present in the repository: interfaces declared there need to be pulled into the generated code.  Fortunately, the `Relative_path.t` data structure that is used to represent paths within the type checker has a prefix component that can be `Root` and `Hhi`, and the latter is used for built-in interface files, while `.php` as well as `.hhi` files from the repository have the prefix `Root`.

Reviewed By: kmeht

Differential Revision: D17980259

fbshipit-source-id: 48e4b1ed9e80c30feeb5acf39e87d9398d24f1a8