Ensure that a pipe is not closed before we write "stop" to it
Summary:
Every SQLite repo that HHVM writes to eventually gets corrupt. I've started collecting corrupt SQLite repos and browsing them - you can see the code I'm using in fbsource commit 51b52d9c8cea61f3563ab976e53856f4430b76db.

In every corrupt DB, SQLite detects corruption inside of `btreeInitPage`. This occurs in `decodeFlags()`, and always because `flagByte` equals 115, which is much larger than all the flags or-ed together could possibly be.

Going up a frame in GDB, we see that this is because 115 is ascii for "s", and that the flag byte has been overwritten with the word "stop\0".

This pattern seems to hold in every corrupt SQLite DB I've seen, regardless of whether it's a SQLFacts, native autoloader, or `.hhbc` repo.

By fbgs-ing for "stop", I've found a plausible explanation. In CLI server mode, workers write "stop" to a pipe to tell the server that they're ready to terminate. But the pipe file descriptor is closed in a different thread from where we write "stop", and we don't have any guarantee that the file descriptor is still safe to write to.

My proposed order of execution:

1. The `m_monitor` enters one of its early returns, closing `m_wpipe` before we've called `MonitorThread::join`
2. A SQLite file is opened. The OS recycles file descriptors, so the new SQLite file receives the same file descriptor `m_wpipe` used to have.
3. We call `MonitorThread::join`, either explicitly or in its destructor. This writes "stop" to the file descriptor, which now points to the SQLite file.

Because we always join our child thread, and because therefore we're guaranteed to outlive our child thread, we can resolve this problem by managing the file descriptors' lifetimes outside of the child thread.

Reviewed By: alexeyt, markw65

Differential Revision: D17970274

fbshipit-source-id: e2e5b8893a33cfcbbe58ede41ba68a672bcdd762