Don't constrain before punting LIterNext
Summary:
If HHBBC doesn't generate AssertCells for the iterator output locals, we'll punt to the interpreter for IterNext.

If we constrain the DataType of the base prior to punting, we'll generate multiple profiling translations for this case and stitch them together. Instead, if we do the constraint afterwards, we'll generate a single InterpOne (because it's generic over DataType). Doing the latter is a significant icache win and a CPU win.

Reviewed By: mofarrell

Differential Revision: D17988647

fbshipit-source-id: 6d7d31fc766c053eac62fd5eb86fa158fc4acff9