Fix overzealous optimization of CombineAndResolveTypeStruct in hhbbc
Summary:
The hhbbc opts for reified generics share type structure code with the `is` and `as` associated bytecodes. We need to fully resolve all generics for reification, while `is` and `as` do not care about the generic parameters of builtin containers (as these are not enforced).

`VerifyParamTypeTS` and `VerifyRetTypeTS` where already indicating that they need the full recursive resolution of all type structures in calls to `resolveTSStatically`. `CombineAndResolveTypeStruct` was incorrectly indicating to `resolveTSStatically` that it did not care about this, additionally the interp step for this bytecode calls `canReduceToDontResolve` which was not taking this into account either.

Reviewed By: oulgen

Differential Revision: D17992182

fbshipit-source-id: 964fdbaf2be09260634449374fbae3b5c9cdffc7