skeleton of main loop
Summary:
This is the skeleton of the main loop of hh_decl.

It includes cachelib initialization, and passing it to the prototype, and getting back the base address.

It doesn't yet include decl requests from the workers.

I tried to keep most thread-launching in a single central place, on the grounds that threads are a precious resource and it'd be nice to see precisely where and why they're spawned. (the only ones that didn't are the ones from the "prototype-watcher", because it got too fiddly otherways).

It's kind of a long way from validating that any of this works! I think the next step is to hook it up to the real prototype, and then send some decl requests. I'll need to figure out a plan for unit tests before then, for sure, because that's too much to write blind.

Differential Revision: D17758913

fbshipit-source-id: c39c57790453ff53331cc0b9c8da8bf951551e6e