Make a field mutable and add message prefixes in thrift serializer
Summary:
Make the `prop` field mutable and pass `FieldInfo` by const reference per discussion in D16821721.

Also add `[Suppressed]` and `[Unsuppressed]` prefixes to warning messages to be able to distinguish the case of the error suppressed by an attribute from an unsuppressed one and from the warnings from old hhvm. This way we can figure out when all the problematic cases have been annotated with the new attribute and we can turn errors on by default.

Reviewed By: alexeyt

Differential Revision: D17866247

fbshipit-source-id: 39fb46f4ca3118a106f29f082f2c0ed260df67fd