Fix issues with no-return memoized functions
Summary:
MemoGet* instructions don't push a value on the miss path.

DCE dealt with properly merging the two paths to get the out-state
during global analysis/optimization, but during local optimization
there was an unhandled edge case.

If we knew that the MemoGet would miss, but couldn't optimize it away
(because it might throw for bad inputs), the out state of the block
wouldn't include the value; this meant that local dce would
incorrectly try to inspect/pop the top of the dce stack when it might
be empty, and global dce might incorrectly assert that the dce stack
didn't match the interp stack.

Fix it so that dce's stack always has the output value, even when we
know it will never produce one.

Reviewed By: ricklavoie

Differential Revision: D17993743

fbshipit-source-id: 94a9b91f692abda96b56e8866987a5ff62c1d7ca