Fix erroneous assert in vasm-graph-color spill hoisting
Summary:
Originally we would only generate a reload before a Vreg's
use. Therefore if we saw a reload for a Vreg within a loop, we
asserted that the Vreg had been marked as used within the loop. This
is no longer true with the more aggressive rematerialization logic, as
we may decide to reload a Vreg at a phi or copy, even if its not
actually used at that point. Therefore we can have reloads of Vregs
that aren't used within the loop.

Remove the assert. The logic (mainly) worked already even with the
reload present. Take it into account when calculating hoisting
profitability.

Reviewed By: mofarrell

Differential Revision: D18002123

fbshipit-source-id: ab9d5934bc1b512b1ff3653991c2f46ff0a27b64