Avoid mutating global state in Instruction_sequence
Summary:
Instruction sequence is at the bottom of the dependency graph
of emitter and should not really call into higher parts emit_*
that are the interface (top-level) parts.  This is particularly
problematic in Rust because the `rewrite_tv` function was
mutating the global emitter state through `emit_adata`.

Therefore, move the state-mutating function to `emit_adata`,
which also enables hiding more of intermediate results,
and rename it to the more readable `rewrite_typed_values`.

Reviewed By: shiqicao

Differential Revision: D17972750

fbshipit-source-id: 57d8d73847c83e86cd32de21f95cdf95f8d42a37