Remove ability to enter TC at prologues
Summary:
This code path allows us to perform the prologue logic in JIT instead of C++.
It is quite challenging to support, as once we defer ActRec setup to the
end of the prologue, we won't have a valid vmfp() when entering a new VM
instance. It is not worth paying the cost of maintaining separate vmfp() from
the parent instance for exception handling and separate native fp for the
purpose of frame linkage.

This is also well aligned with the goal of reducing and eventually eliminating
VM reentrancy.

Reviewed By: ricklavoie

Differential Revision: D17961977

fbshipit-source-id: aeb3da404c2a0b4f5ad0c4380a7bb4d39757462c