Defer ActRec materialization to the callee
Summary:
Instead of writing an ActRec in a Call IR opcode, set up a set of registers
containing information necessary to write an ActRec, call the target function's
prologue and write the ActRec there.

This also allows various function binding/dispatch stubs to use these same
registers instead of loading from an ActRec.

When writing an ActRec in a prologue of a closure, load the context from the
closure and avoid storing it again via InitCtx.

This diff splits the work of setting up each ActRec field from one complex
instruction to two simpler instructions, so it increases the overall number of
retired instructions. However, it reduces the amount of code at each caller
significantly, reducing CPU time usage.

Reviewed By: ricklavoie

Differential Revision: D17977166

fbshipit-source-id: c9e39c14b6e7253d047178fa815fd5db22a80f05