Restructure the hhbbc parser
Summary:
This lets us produce hhbbc Units as we produce each UnitEmitter. This
is a very small performance improvement by itself, but it also opens
the way to building parts of the index in parallel with the parse
step. Finally, since the caller determines what happens to the
UnitEmitter (rather than handing them all off in a vector of
unique_ptrs) it should make it possible to keep writing to the cache
(when configured) in the background while hhbbc runs.

Reviewed By: ricklavoie

Differential Revision: D17686959

fbshipit-source-id: 0419f74a124fd93194ad1eeb4f79044aec73db4d