Inline MakeStructDArray and friends
Summary:
A "struct array" is an array with string literal keys (but not with constant values - that would just make it static).

We construct these arrays by spilling the values to the stack and then calling MakeStructDArray (and related helpers). There are three ways we waste compute in this helper:
- The plurality of the cost (close to 50%) is in dcache misses from loading the hash values of the string keys.
- We pay a cost to hash and store the new string keys.
- We pay a (hidden) cost to store the values to the stack and load them from there again.

We can eliminate all of these costs if we outline the stores to the MixedArray elements into the JIT - we'll burn the hash values of the string keys into the JIT, and store-elim the stack stores.

This diff will increase code size in the TC, but only by two instructions per MixedArray element. The store of the value's type and data fields replace the instructions that store those to the stack. The two extra instructions are the store of the key and its hash.

Reviewed By: ricklavoie

Differential Revision: D17987597

fbshipit-source-id: 46d8ad0e050a41746c9c5f12298aff076a64d49f