Avoid expensive sorting when choosing a physical register
Summary:
In vasm-graph-color, when selecting the best physical register for a
particular Vreg, we calculate all of the penalty weights for each
physical register, and then sort the registers by them. However, this
is expensive, mainly because it's done so much. In the common case,
the "best" physical register is available. So, we just want the first
register in the sorted list. We only need the fully sorted list in the
case where the best register isn't available, and we need to displace
other Vregs.

Instead of always sorting the registers, first just do a linear walk
over the registers, keeping track of the best one. If the best one is
available, then we're done and return it. If not, *then* do the full
sort and use the existing logic.

Reviewed By: jano

Differential Revision: D18017605

fbshipit-source-id: a16d4f5b70c090a7a200a09708057e4576fd0b8b