Speed up computePreds()
Summary:
computePreds() was returning its results in a vector of vectors, which
is inefficient. The majority of blocks has less than 2 predecessors,
so we're allocating lots of little pieces of memory. Instead use a
vector of TinyVectors. This can still handle arbitrary number of
predecessors, but is way more optimized for the common case.

Reviewed By: jano

Differential Revision: D18019724

fbshipit-source-id: 95fe686ae23273983554a2cfe4666105119912c6