Avoid unnecessary compute_rpo() call in vasm-graph-color
Summary:
After collapsing split side exits back, we call compute_rpo() to
recalculate the RPO ordering for the unit (since splitting and
collapsing the side exits may have changed it). However, nothing
actually uses this RPO information, unless we call
calculate_sp_offsets(), and we only do that if we're in a debug
build. So likewise only recalculate RPO in debug builds.

Reviewed By: jano

Differential Revision: D18019870

fbshipit-source-id: 12456d1e0257483ec263052c67b576500bdf4304