Optimize spill liveness algorithm in vasm-graph-color
Summary:
When calculating the liveness of spill slots, the liveness is
represented by a vector of dynamic_bitsets. One dynamic_bitset for
each block, with each bit in the bitset representing a spill
slot. However, a vector of dynamic_bitsets is a very heavyweight
representation and is rather expensive to allocate, even if there's
very few operations performed on them. Use std::bitset would be
faster, as it's a fixed-size value-type (so no memory
allocation). However we cannot just use std::bitset because we don't
have an intrinsic upper bound on the number of possible spill slots
(and some units can have hundreds).

So, templatize the logic and conditionally use one representation or
the other. If the number of spill slots for a particular unit is below
a threshold, just use a fixed size std::bitset. Otherwise use a
dynamic_bitset as normal. The threshold has been chosen as 128 slots
as that seemed to give the best results.

Reviewed By: jano

Differential Revision: D18018754

fbshipit-source-id: 8e0c6ec94e8d64827666a731de6659c8382fc831