Fix native stack imbalance when fcallHelper throws
Summary:
FCallHelperThunk was converted from phplogue to stublogue, which moves native
stack by 16 bytes, but the diff failed to add a corresponding code to move it
back in case of an exception. This was confusing unwinder, making it not call
the caller's catch block.

This code path is used only very rarely when we fail to obtain a write lease.

Reviewed By: ricklavoie

Differential Revision: D18031201

fbshipit-source-id: 67d5bb174a86f45e963db8180e850ba304ad1710