Pack extra arguments before entering N+1 prologue
Summary:
Move the responsibility of packing the extra arguments from prologues to either
emitter (known callees) or redispatch stub (unknown callees).

Simplifies things a bit and makes it possible to treat input arguments in
prologue as a stack rather than as a locals on a frame, as the stack layout
is now statically known. This is important for being able to defer ActRec
spill further down.

Reviewed By: ricklavoie

Differential Revision: D18025354

fbshipit-source-id: 896f066614a2796901194b107295b662542f1ab0