Strip HH\ prefix for autoimported names in Errors
Summary:
This updates `Errors` to strip the `HH\` prefix from autoimported names.

We do this because some extremely common names are autoimported, like Collections and Awaitables. As we unify the typechecker and compiler's understanding of autoimported types, we don't want the `HH\` to appear in error messages and hover types, because it may confuse users into writing that in code.

Differential Revision: D17987427

fbshipit-source-id: c5a2627b413cfefd29969acc21b484be7cf92f8d