Remove plain array use in ext_soap
Summary:
The second arg to `self::cleanArrays` is a "base" container into which the values of the first argument are put and then this modified base is returned. It's not a big deal since it gets cast to varray anyway and so doesn't escape to WWW, but I saw it, so let's just clean it up.

It needs to be a darray here instead of a varray, because you can't implicitly append to a vec.

Reviewed By: billf

Differential Revision: D18019540

fbshipit-source-id: 3b0db28c734a54bbd34ce34373c058740efa1d92