Do not pessimize Shapes:: arguments
Summary: This diff effectively undoes {D17108480}. The `Shapes::at` and `Shapes::idx` functions are special cased in the type checker and their arguments needed to be manually pessimized. This removes the manual pessimization of the arguments and instead calls coercion, so the calls have the same behavior as normal function calls.

Reviewed By: Matt-Schellhas

Differential Revision: D17698561

fbshipit-source-id: 903c75737031f514c9eaba8640ddd4060f0aceed