Move toplevel ns elaboration to aast only
Summary:
This diff is part of the effort to get rid of the legacy AST as well as to help the Rust port.

Currently, we perform toplevel namespace elaboration on the AST before we convert it to an AAST.

This diff changes the elaboration pass to work on the AAST instead of the legacy AST and replaces the AST version. This should help unblock deleting the legacy AST.

Reviewed By: shiqicao

Differential Revision: D18048594

fbshipit-source-id: c7093848e85db05f19d8abde43b703b6149c0e4d