Split DefSP into DefFrameRelSP and DefRegSP
Summary:
DefSP works differently in resumed vs non-resumed functions. Split it into
these two use cases and determine which opcode to use at the IRGS init time.

Remove DefCallFP, as it is identical to DefRegSP. Prologues now set up their SP
twice, first in the context of the caller using DefRegSP, then in the context
of the callee using DefFrameRelSP.

Allows work to be moved before DefFuncEntryFP and operate on a stack rather
than locals.

Reviewed By: ricklavoie

Differential Revision: D18046737

fbshipit-source-id: 67f8a9ce0fb80e3eeb82c24eb440fcb612e7575d