Ignore Box in variants with a single boxed tuple argument
Summary:
In Rust, even nullary variants of an enum occupy just as much space as its largest variant. Because of this, the aast::Expr_ occupies a lot of space in memory as currently written. {D17772799} will address this by wrapping the fields of its variants in boxed tuples. Adding a box will reduce all variants of Expr_ to a much more reasonable size (as currently implemented, D17772799 shrinks them from 712 bytes to 32).

{D17696666} has the procedural macro which derives the Ocamlvalue trait ignore the box when converting a variant with a single field which is a boxed tuple. Instead, it directly places those fields inline in the OCaml representation of the variant. This diff does the same for `ocamlrep`.

Reviewed By: shiqicao

Differential Revision: D17864959

fbshipit-source-id: a231d22f7203f9da26fa441139c8edbc82e3f695