Borrow Arena mutably in allocation methods and return *-ptr
Summary:
For performance reasons, we would like to borrow Arenas mutably (to get rid of its use of RefCell) and stop bounds-checking writes to block fields.

If we borrow Arenas mutably in allocation methods, and we associate the block-initializer with the liftetime of the allocation borrow (as we previously did with BlockBuilder), then we cannot convert our fields while the block-initializer is in scope. This is because only one mutable borrow of a value is permitted at a time, so we cannot allocate (mutating the arena) while the BlockBuilder (which has permission to mutate the arena) exists. This requires us to convert our fields ahead of time, write their values to the stack, and *only then* allocate the block and copy the stack-allocated values into it. This is the approach taken by ocamlpool_rust::utils and {D17864962}.

However, this approach uses a lot of stack space. We could avoid doing so if we didn't associate the block-initializer type with the mutable borrow of the Arena. This diff does this by making that block-initializer type a `*mut Value<'a>` rather than a `BlockBuilder` (which contains a `&mut [Value<'a>]`.

Since this diff removes BlockBuilder, it seems like a good time to also move field-writing into Arena (so that it can be made an Allocator method in {D17864960}, allowing us to use `caml_set_field` for the ocamlpool backend in {D17864957}). Since the block-initializer type is now just a pointer to the first field, these writes are no longer bounds-checked, so the `set_field` method must be marked `unsafe`. We could make it safe by returning a `(*mut Value, usize)` pair from allocation methods, communicating the size, or by reading the size from the block header, but I'm not sure this is necessary. Not doing an out-of-bounds write here doesn't seem so difficult, and most implementations of `into_ocamlrep` are derived by the ocamlrep procedural macro rather than hand-written anyway.

Reviewed By: ljw1004

Differential Revision: D18008608

fbshipit-source-id: 946f436a40866b0072918721598e047e72d42f6c